bugfix#826: refacto search-input & create base-textfield
