+++ libs/web-components/search-input/README.md
@@ -1,10 +1,12 @@
 # Search Input
+
 [![See it on NPM!](https://img.shields.io/npm/v/@finastra/search-input?style=for-the-badge)](https://www.npmjs.com/package/@finastra/search-input)
 [![How big is this package in your project?](https://img.shields.io/bundlephobia/minzip/@finastra/search-input?style=for-the-badge)](https://bundlephobia.com/result?p=@finastra/search-input)
 [![Storybook](https://shields.io/badge/-Play%20with%20this%20web%20component-2a0481?logo=storybook&style=for-the-badge)](https://finastra.github.io/finastra-design-system/?path=/story/forms-search-input--default-story)
 
-`<fds-search-input>` delivers a single input field with a "reset" button as well as a mangifying glass icon with which to power search interactions.
+`<fds-search-input>` delivers a single input field with a "reset" button as well as a loader and a mangifying glass icon with which to power search interactions.
 
+Search Input Component extends `fds-base-textfield` component, so it inherits all capacity of [fds-textfield](https://github.com/Finastra/finastra-design-system/blob/master/libs/web-components/textfield/README.md)
 
 ## Usage
