+++ libs/web-components/search-input/demo/index.html
@@ -1,7 +1,21 @@
 <script type="module" src="node_modules/@finastra/search-input/dist/src/search-input.js"></script>
 
-Search with Label
-<fds-search-input label="Search"></fds-search-input>
+<style>
+  .demo-title {
+    font: var(--fds-headline-5, #FFFFFF);
+    color: var(--fds-on-surface, normal normal var(--fds-headline-5-font-weight, 800) var(--fds-headline-5-font-size, 1rem)/var(--fds-headline-5-line-height, 1.375rem) var(--fds-headline-5-font-family, Spartan));
+    margin: var(--fds-spacing-2, 8px) var(--fds-spacing-0, 0px);
+  }
+</style>
 
-Search with Label + loading
-<fds-search-input label="Search" loading="true"></fds-search-input>
+<div style="display: flex; flex-direction: column; gap: 24px;">
+  <div>
+    <div class="demo-title">Search</div>
+    <fds-search-input placeholder="Search ..." value="Banking"></fds-search-input>
+  </div>
+
+  <div>
+    <div class="demo-title">Search with loader</div>
+    <fds-search-input value="loading ..." loading disabled></fds-search-input>
+  </div>
+</div>
