--- libs/web-components/search-input/stories/search-input.stories.mdx
@@ -1,101 +1,0 @@
-import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs';
-import { ifDefined } from 'lit/directives/if-defined.js';
-import { html } from 'lit';
-import '@finastra/search-input';
-
-<Meta
-  title="FORMS/Search Input"
-  component={'fds-search-input'}
-  argTypes={{
-    icon: {
-      table: {
-        category: 'properties',
-        defaultValue: { summary: 'search' }
-      },
-      control: 'text',
-      defaultValue: 'search',
-      type: 'string',
-      description: 'icon to display in the search input'
-    },
-    showClearButton: {
-      table: {
-        category: 'properties',
-        defaultValue: { summary: true }
-      },
-      control: 'boolean',
-      defaultValue: true,
-      type: 'boolean',
-      description: 'show clear button'
-    },
-    loading: {
-      table: {
-        category: 'properties',
-        defaultValue: { summary: false }
-      },
-      control: 'boolean',
-      defaultValue: true,
-      type: 'boolean',
-      description: 'show loading button'
-    }
-  }}
-/>
-
-export const Template = ({ icon, showClearButton }) => {
-  return html`<fds-search-input .icon=${ifDefined(icon)} .showClearButton=${ifDefined(showClearButton)}></fds-search-input>`;
-};
-
-# Search Input
-
-Search Input Component is a Text Input field with clear button and search icon.
-
-Search Input extends `mwc-textfield` component, so it inherits all capacity of [mwc-textfield](https://github.com/material-components/material-components-web-components/tree/master/packages/textfield)
-
-<Canvas>
-  <Story
-    name="Default"
-    parameters={{
-      docs: {
-        source: {
-          code: '<fds-search-input></fds-search-input>'
-        }
-      },
-      design: {
-        type: 'figma',
-        url: 'https://www.figma.com/file/E1Mb1556RT3HbAUVu2Q0LV/Finastra-design-system?node-id=56704%3A19089'
-      }
-    }}
-  >
-    {Template.bind({})}
-  </Story>
-</Canvas>
-
-## Installation
-
-```
-npm install @finastra/search-input
-
-```
-
-## API
-
-<ArgsTable of="." story="^" />
-
-## Examples
-
-### Custom icon
-
-<Canvas>
-  <Story
-    name="Custom Icon"
-    args={{ icon: 'location_searching' }}
-    parameters={{
-      docs: {
-        source: {
-          code: '<fds-search-input icon="location_searching"></fds-search-input>'
-        }
-      }
-    }}
-  >
-    {Template.bind({})}
-  </Story>
-</Canvas>
