+++ libs/web-components/textfield/README.md
@@ -1,4 +1,4 @@
-# Textfield
+# TextField
 
 [![See it on NPM!](https://img.shields.io/npm/v/@finastra/textfield?style=for-the-badge)](https://www.npmjs.com/package/@finastra/textfield)
 [![How big is this package in your project?](https://img.shields.io/bundlephobia/minzip/@finastra/textfield?style=for-the-badge)](https://bundlephobia.com/result?p=@finastra/textfield')
@@ -34,7 +34,7 @@ import '@finastra/textfield';
 
 Date pickers and Time pickers allow selecting a value from a standard format.
 
-The **Textfield** support date `type="date"`, time `type="time"` and date&time `type="datetime-local"` pickers. See [MDN web docs](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/date) for more information about the different `input` types.
+The **TextField** support date `type="date"`, time `type="time"` and date&time `type="datetime-local"` pickers. See [MDN web docs](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/date) for more information about the different `input` types.
 
 #### Example
