+++ libs/web-components/textfield/demo/index.html
@@ -1,7 +1,7 @@
-<script type="module" src="node_modules/@finastra/textfield/dist/src/textfield.js"></script>
+<script type="module" src="node_modules/@finastra/textfield/dist/src/index.js"></script>
 <script type="module" src="node_modules/@finastra/icon-button/dist/src/icon-button.js"></script>
 
-<div>
+<div style="display: flex; flex-direction: column; gap: 24px;">
   <fds-textfield type="password" showActionButton id="my-textfield-dense" dense>
     <fds-icon-button slot="actionButton" icon="visibility_off" onclick="myFunction(this)"></fds-icon-button>
   </fds-textfield>
