+++ libs/web-components/textfield/package.json
@@ -1,16 +1,16 @@
 {
   "name": "@finastra/textfield",
-  "displayName": "Textfield",
+  "displayName": "TextField",
   "version": "0.0.39",
-  "description": "Textfield Web Component",
+  "description": "TextField Web Component",
   "keywords": [
     "Finastra Design System",
     "Web Components",
-    "Textfield"
+    "TextField"
   ],
   "license": "MIT",
-  "main": "dist/src/textfield.js",
-  "module": "dist/src/textfield.js",
+  "main": "dist/src/index.js",
+  "module": "dist/src/index.js",
   "scripts": {
     "build": "webpack",
     "build:style": "node ../../../scripts/sass-to-lit-css/index.js src/styles.scss"
