+++ libs/web-components/textfield/src/index.ts
@@ -1,0 +1,3 @@
+export * from './base-textfield';
+export * from './textfield';
+
