+++ libs/web-components/textfield/test/textfield.test.ts
@@ -1,10 +1,10 @@
 import { html, fixture, expect, elementUpdated } from '@open-wc/testing';
-import { Textfield } from '../src/textfield.js';
+import { TextField } from '../src/textfield.js';
 import '../src/textfield.js';
 
-describe('Textfield', () => {
+describe('TextField', () => {
   it('loads accessibly', async () => {
-    const el: Textfield = await fixture(html`<fds-textfield label='test'></fds-textfield>`);
+    const el: TextField = await fixture(html`<fds-textfield label='test'></fds-textfield>`);
 
     await elementUpdated(el);
     await expect(el).to.be.accessible();
